文件删除校验

+++ app/Http/Controllers/InstagramController.php
@@ -203,7 +203,9 @@ class InstagramController
 
         $img = file_get_contents($imgUrl);
 
-        $fileName = date('YmdHis') . rand(1, 99) . '.' . $suffix;
+        $path = '/usr/local/src/images/';
+
+        $fileName = $path . date('YmdHis') . rand(1, 99) . '.' . $suffix;
 
         file_put_contents($fileName, $img);
 
@@ -226,7 +228,9 @@ class InstagramController
 
         $imgObj->uploadImage($key, $filePath);
 
-        unlink($filePath);
+        if (file_exists($filePath)) {
+            unlink($filePath);
+        }
 
         return $key;
     }
